docs: use theming parameter in App stories
